pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "docker-janitor"
version = "0.1.0"
description = "Automated Docker resource cleanup with configurable policies."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "docker>=7.0",
    "pyyaml>=6.0",
    "requests>=2.31",
]

[project.scripts]
docker-janitor = "janitor.cli:cli"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
    "types-PyYAML",
    "types-requests",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
